@@ -40,14 +40,14 @@
     "update-ns-webpack": "update-ns-webpack"
   },
   "dependencies": {
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/http": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/forms": "8.0.0",
+    "@angular/http": "8.0.0-beta.10",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/router": "8.0.0",
     "email-validator": "^2.0.4",
     "nativescript-angular": "~7.2.0",
     "nativescript-iqkeyboardmanager": "~1.3.0",
@@ -59,8 +59,8 @@
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "~7.1.0",
-    "@ngtools/webpack": "~7.1.0",
+    "@angular/compiler-cli": "8.0.0",
+    "@ngtools/webpack": "8.0.0",
     "@types/chai": "~4.1.7",
     "@types/mocha": "~5.2.5",
     "@types/node": "~10.12.18",
@@ -81,6 +81,6 @@
     "nativescript-dev-typescript": "~0.7.9",
     "nativescript-dev-webpack": "~0.19.1",
     "tslint": "^5.4.2",
-    "typescript": "~3.1.1"
+    "typescript": "~3.4.5"
   }
 }
